키워드 알림 서비스 추가

+++ src/main/java/in/koala/mapper/CrawlingMapper.java
@@ -16,7 +16,7 @@ public interface CrawlingMapper {
 
     int checkDuplicatedData(Crawling crawling);
 
-    int updateLog(Short site, Timestamp crawlingAt);
+    int updateLog(Integer site, Timestamp crawlingAt);
 
     void addCrawlingData(List<Crawling> crawlingInsertList);
