+++ src/main/java/in/koala/service/CrawlingService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface CrawlingService {
 
     String test();
-    void updateLog(Short site, Timestamp crawlingAt);
+    void updateLog(Integer site, Timestamp crawlingAt);
     void updateTable(List<Crawling> crawlingInsertList, List<Crawling> crawlingUpdateList);
     Timestamp getLatelyCrawlingTime();
